safekeeper: implement wal reader fanout
The basic idea for wal reader fanout is that wal decoding and
interpretation happens in a separate tokio task and senders
attach to it. Senders only receive batches concerning their shard
and only past the Lsn they've last seen.

When the wal reading task is spawned, it returns a reference through
which it can be controlled. This reference is shared by multiple wal
senders. When all wal senders go out of scope, the reference goes out of
scope and the task is aborted.