Don't include sources for extensions with no runnable tests
The run_tests.sh script tries to run "neon-test.sh" or "make
installcheck" for every directory in /ext-src. Hide the directories
for extensions that we don't want to test. Or in case of
onnxbuild-src, have a no-op "neon-test.sh" file to silence the error
that you get on "make installcheck" in that directory.